+++ examples/sqrt_lut.py
@@ -1,5 +1,5 @@
 """
-Creates a blueprint that calculates the sqrt of `signal-I` and gives the output 
+Creates a blueprint that calculates the sqrt of `signal-I` and gives the output
 on `signal-O` with a time complexity of 1 tick. Handles the entire positive
 integer range.
 
@@ -21,12 +21,19 @@ try:
 except ImportError:
     pyperclip = None
 
+
 def main():
     bp = Blueprint()
 
     unique_signals = []
     for signal_name in signals.raw:
-        if signal_name in {"signal-item-parameter", "signal-fuel-parameter", "signal-fluid-parameter", "signal-signal-parameter", "signal-unknown"}:
+        if signal_name in {
+            "signal-item-parameter",
+            "signal-fuel-parameter",
+            "signal-fluid-parameter",
+            "signal-signal-parameter",
+            "signal-unknown",
+        }:
             continue
         if signal_name in signals.pure_virtual:
             continue
@@ -37,25 +44,32 @@ def main():
     def add_row(dc, cc):
         bp.entities.append(cc, id=f"cc_{num_ccs}", tile_position=(0, num_ccs))
         bp.entities.append(dc, id=f"dc_{num_ccs}", tile_position=(1, num_ccs))
-        bp.add_circuit_connection("green", entity_1=f"cc_{num_ccs}", entity_2=f"dc_{num_ccs}")
+        bp.add_circuit_connection(
+            "green", entity_1=f"cc_{num_ccs}", entity_2=f"dc_{num_ccs}"
+        )
         try:
-            bp.add_circuit_connection("red", entity_1=f"dc_{num_ccs - 1}", entity_2=f"dc_{num_ccs}")
-            bp.add_circuit_connection("red", entity_1=f"dc_{num_ccs - 1}", side_1="output", entity_2=f"dc_{num_ccs}", side_2="output")
+            bp.add_circuit_connection(
+                "red", entity_1=f"dc_{num_ccs - 1}", entity_2=f"dc_{num_ccs}"
+            )
+            bp.add_circuit_connection(
+                "red",
+                entity_1=f"dc_{num_ccs - 1}",
+                side_1="output",
+                entity_2=f"dc_{num_ccs}",
+                side_2="output",
+            )
         except KeyError:
             pass
 
     import math
+
     num_squares = math.floor(math.sqrt(2**31 - 1))
 
     dc = DeciderCombinator("decider-combinator", direction=Direction.EAST)
     Input = DeciderCombinator.Input
     Output = DeciderCombinator.Output
-    dc.conditions = [
-        Input("signal-each", {"green"}) <= Input("signal-I", {"red"})
-    ]
-    dc.outputs = [
-        Output("signal-O", copy_count_from_input=False)
-    ]
+    dc.conditions = [Input("signal-each", {"green"}) <= Input("signal-I", {"red"})]
+    dc.outputs = [Output("signal-O", copy_count_from_input=False)]
 
     num_signals = len(unique_signals)
     cc = ConstantCombinator("constant-combinator", direction=Direction.EAST)
@@ -94,6 +108,7 @@ def main():
         with open("examples/output/sqrt_lut.blueprint", "w") as file:
             file.write(bp.to_string())
         print("Wrote 'examples/output/sqrt_lut.blueprint'")
+
 
 if __name__ == "__main__":
     main()